+++ src/main/java/com/kinoxp/repository/MovieRepository.java
@@ -3,4 +3,4 @@ package com.kinoxp.repository;
 import com.kinoxp.model.movie.Movie;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface MovieRepository extends JpaRepository<Movie, Integer> {}
+public interface MovieRepository extends JpaRepository<Movie, Long> {}
